Abstract websockets into an interface for easy switching

+++ DiscordSharp/Events/DiscordSocketClosedEventArgs.cs
@@ -4,7 +4,7 @@ namespace DiscordSharp
 {
     public class DiscordSocketClosedEventArgs : EventArgs
     {
-        public ushort Code { get; internal set; }
+        public int Code { get; internal set; }
         public string Reason { get; internal set; }
         public bool WasClean { get; internal set; }
